Feat docker admin (#31)
* Added Dockerfile Updated Admin

* Joke api integrated

* Joke api integrated

* Migrations issues fixed (i hope)

* Migrations issue solved

* updated routes

* update-routes-migrations

---------

+++ database/migrations/2024_09_13_003745_create_orders_table.php
@@ -6,24 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('orders', function (Blueprint $table) {
-            $table->id();
-            $table->date('creationDate');
-            $table->date('deliveryDate');
-            $table->integer('totalPrice');
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('orders')) {
+            Schema::create('orders', function (Blueprint $table) {
+                $table->id();
+                $table->date('creationDate');
+                $table->date('deliveryDate');
+                $table->integer('totalPrice');
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
+
     public function down(): void
     {
         Schema::dropIfExists('orders');
